Refactor: Cars 일급 컬렉션으로 포장
RacingGame 클래스에 있던 List<Car> cars 멤버 변수를 Cars 클래스로 분리하여 외부에서 추가하거나 수정할 수 없도록 일급컬렉션으로 포장하였습니다.
